+++ sysarq/src/tests/Documents/Create/CreateAdministrativeProcess.test.js
@@ -1,7 +1,13 @@
 import { render, screen, fireEvent, within } from "@testing-library/react";
 
 import { server } from "../../support/server";
-import { input, submitClick } from "../../support";
+import {
+	input,
+	submitClick,
+	abbreviationSelector,
+	shelfSelector,
+	rackSelector,
+} from "../../support";
 
 import CreateAdministrativeProcess from "../../../pages/Documents/Create/CreateAdministrativeProcess";
 
@@ -34,13 +40,16 @@ describe("Create Administrative Process Screen Test", () => {
 		expect(screen.getByText(REQUIRED_DATE_ERROR_MESSAGE)).toBeInTheDocument();
 
 		input(NOTICE_DATE_LABEL, "01/02/");
+
 		expect(
 			screen.queryByText(REQUIRED_DATE_ERROR_MESSAGE)
 		).not.toBeInTheDocument();
+
 		submitClick();
 		expect(screen.getByText(INVALID_DATE_ERROR_MESSAGE)).toBeInTheDocument();
 
 		input(NOTICE_DATE_LABEL, "03/04/2005");
+
 		expect(
 			screen.queryByText(INVALID_DATE_ERROR_MESSAGE)
 		).not.toBeInTheDocument();
@@ -50,13 +59,16 @@ describe("Create Administrative Process Screen Test", () => {
 		expect(screen.getByText(REQUIRED_DATE_ERROR_MESSAGE)).toBeInTheDocument();
 
 		input(ARCHIVING_DATE_LABEL, "36/07/2008");
+
 		expect(
 			screen.queryByText(REQUIRED_DATE_ERROR_MESSAGE)
 		).not.toBeInTheDocument();
+
 		submitClick();
 		expect(screen.getByText(INVALID_DATE_ERROR_MESSAGE)).toBeInTheDocument();
 
 		input(ARCHIVING_DATE_LABEL, "09/10/2011");
+
 		expect(
 			screen.queryByText(INVALID_DATE_ERROR_MESSAGE)
 		).not.toBeInTheDocument();
@@ -66,6 +78,7 @@ describe("Create Administrative Process Screen Test", () => {
 		expect(screen.getByText("Insira um período válido")).toBeInTheDocument();
 
 		input("Referência", "04/2015");
+
 		expect(
 			screen.queryByText("Insira um período válido")
 		).not.toBeInTheDocument();
@@ -74,6 +87,7 @@ describe("Create Administrative Process Screen Test", () => {
 		expect(screen.getByText("Insira o número do processo")).toBeInTheDocument();
 
 		input("Número do Processo*", "16");
+
 		expect(
 			screen.queryByText("Insira o número do processo")
 		).not.toBeInTheDocument();
@@ -83,16 +97,20 @@ describe("Create Administrative Process Screen Test", () => {
 		expect(screen.getByText("Insira somente números")).toBeInTheDocument();
 
 		input("CPF/CNPJ", "2324252627");
+
 		expect(
 			screen.queryByText("Insira somente números")
 		).not.toBeInTheDocument();
 		submitClick();
+
 		expect(screen.getByText("Insira um CPF/CNPJ válido")).toBeInTheDocument();
 
 		input("CPF/CNPJ", "28293031323");
+
 		expect(
 			screen.queryByText("Insira um CPF/CNPJ válido")
 		).not.toBeInTheDocument();
+
 		submitClick();
 		expect(screen.getByText("Insira um interessado")).toBeInTheDocument();
 
@@ -146,20 +164,11 @@ describe("Create Administrative Process Screen Test", () => {
 
 		input("Servidor que Encaminhou", "sender_worker_test");
 
-		fireEvent.mouseDown(screen.getByLabelText("Sigla da Caixa"));
-		const abbreviationOptions = within(screen.getByRole("listbox"));
-		await abbreviationOptions.findByText("abbreviation_test");
-		fireEvent.click(abbreviationOptions.getByText(/abbreviation_test/i));
+		await abbreviationSelector();
 
-		fireEvent.mouseDown(screen.getByLabelText("Estante"));
-		const shelfOptions = within(screen.getByRole("listbox"));
-		await shelfOptions.findByText("47");
-		fireEvent.click(shelfOptions.getByText(/47/i));
+		await shelfSelector();
 
-		fireEvent.mouseDown(screen.getByLabelText("Prateleira"));
-		const rackOptions = within(screen.getByRole("listbox"));
-		await rackOptions.findByText("49");
-		fireEvent.click(rackOptions.getByText(/49/i));
+		await rackSelector();
 
 		expect(
 			screen.queryByText(UNARCHIVE_DESTINATION_UNIT_LABEL)
@@ -167,6 +176,7 @@ describe("Create Administrative Process Screen Test", () => {
 		expect(
 			screen.queryByText(UNARCHIVE_PROCESS_NUMBER_LABEL)
 		).not.toBeInTheDocument();
+
 		expect(screen.queryByText(UNARCHIVE_DATE_LABEL)).not.toBeInTheDocument();
 
 		fireEvent.mouseDown(screen.getByLabelText("Status*"));
@@ -179,6 +189,7 @@ describe("Create Administrative Process Screen Test", () => {
 		expect(
 			screen.getByText(UNARCHIVE_PROCESS_NUMBER_LABEL)
 		).toBeInTheDocument();
+
 		expect(screen.getByText(UNARCHIVE_DATE_LABEL)).toBeInTheDocument();
 
 		fireEvent.mouseDown(
@@ -201,6 +212,7 @@ describe("Create Administrative Process Screen Test", () => {
 		expect(screen.getByText(INVALID_DATE_ERROR_MESSAGE)).toBeInTheDocument();
 
 		input(UNARCHIVE_DATE_LABEL, "08/09/2055");
+
 		expect(
 			screen.queryByText(INVALID_DATE_ERROR_MESSAGE)
 		).not.toBeInTheDocument();
